# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/memberships": {
            "get": {
                "operationId": "MembershipsController_findAll",
                "summary": "Get all memberships",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MembershipDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "memberships"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "post": {
                "operationId": "MembershipsController_create",
                "summary": "Create new membership",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMembershipDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MembershipDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "memberships"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/memberships/{id}": {
            "patch": {
                "operationId": "MembershipsController_updateMembership",
                "summary": "Update membership",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMembershipDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MembershipDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "memberships"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/api/members/{id}": {
            "get": {
                "operationId": "MembersController_findById",
                "summary": "Get full info about member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "patch": {
                "operationId": "MembersController_update",
                "summary": "Update member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMemberDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/api/acs-keys/member/{memberId}": {
            "get": {
                "operationId": "ACSKeysController_findAllByMemberId",
                "summary": "Get ACS keys for specific member",
                "parameters": [
                    {
                        "name": "memberId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ACSKeyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "acs-keys"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
        "/api/membership-subscriptions/member/{memberId}": {
            "get": {
                "operationId": "MembershipSubscriptionsController_findAllByMemberId",
                "summary": "Get all membership subscriptions for member",
                "parameters": [
                    {
                        "name": "memberId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MembershipSubscriptionDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "membership-subscriptions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/membership-subscriptions/membership/{membershipId}": {
            "get": {
                "operationId": "MembershipSubscriptionsController_findAllByMembershipId",
                "summary": "Get all membership subscriptions for membership",
                "parameters": [
                    {
                        "name": "membershipId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MembershipSubscriptionDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "membership-subscriptions"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "RE: Swynca",
        "description": "B4CKSP4CE member/money management system",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
            "TelegramMetadataDTO": {
                "type": "object",
                "properties": {
                    "telegramId": {
                        "type": "string"
                    },
                    "telegramName": {
                        "type": "string"
                    }
                },
                "required": [
                    "telegramId",
                    "telegramName"
                ]
            },
            "GitHubMetadataDTO": {
                "type": "object",
                "properties": {
                    "githubId": {
                        "type": "string"
                    },
                    "githubUsername": {
                        "type": "string"
                    }
                },
                "required": [
                    "githubId",
                    "githubUsername"
                ]
            },
            "MemberDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "frozen"
                        ]
                    },
                    "balance": {
                        "type": "string"
                    },
                    "joinedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "telegramMetadata": {
                        "$ref": "#/components/schemas/TelegramMetadataDTO"
                    },
                    "githubMetadata": {
                        "$ref": "#/components/schemas/GitHubMetadataDTO"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "status",
                    "balance",
                    "joinedAt"
                ]
            },
...
            "ErrorApiResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
            "MembershipDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "amount",
                    "active"
                ]
            },
            "CreateUpdateMembershipDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "title",
                    "amount",
                    "active"
                ]
            },
            "CreateUpdateMemberDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "required": [
                    "name",
                    "email"
                ]
            },
...
            "ACSKeyDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "memberId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "pan",
                            "uid"
                        ]
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "memberId",
                    "type",
                    "key",
                    "name"
                ]
            },
...
            "MembershipSubscriptionDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "memberId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "membershipId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "subscribedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "declinedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "memberId",
                    "membershipId",
                    "subscribedAt"
                ]
            },
...
        }
    }
}